# api/docs/swagger.yaml
basePath: /
definitions:
  apidocs.ConfigData:
    properties:
      audio_codecs:
        items:
          $ref: '#/definitions/domain.AudioCodec'
        type: array
      global_config:
        $ref: '#/definitions/domain.GlobalConfig'
      hw_accels:
        items:
          $ref: '#/definitions/domain.HWAccel'
        type: array
      output_protocols:
        items:
          type: string
        type: array
      ports:
        $ref: '#/definitions/apidocs.ConfigPorts'
      stream_statuses:
        items:
          $ref: '#/definitions/domain.StreamStatus'
        type: array
      version:
        $ref: '#/definitions/version.Info'
      video_codecs:
        items:
          $ref: '#/definitions/domain.VideoCodec'
        type: array
      watermark_positions:
        items:
          $ref: '#/definitions/domain.WatermarkPosition'
        type: array
      watermark_types:
        items:
          $ref: '#/definitions/domain.WatermarkType'
        type: array
    type: object
...
  apidocs.ConfigUpdateResponse:
    properties:
      global_config:
        $ref: '#/definitions/domain.GlobalConfig'
      ports:
        $ref: '#/definitions/apidocs.ConfigPorts'
    type: object
  apidocs.ErrorBody:
    properties:
      error:
        properties:
          code:
            type: string
          message:
            type: string
        type: object
    type: object
  apidocs.ErrorResponse:
    properties:
      error:
        properties:
          code:
            type: string
          message:
            type: string
        type: object
    type: object
  apidocs.HookData:
    properties:
      data:
        $ref: '#/definitions/domain.Hook'
    type: object
  apidocs.HookList:
    properties:
      data:
        items:
          $ref: '#/definitions/domain.Hook'
        type: array
      total:
        type: integer
    type: object
...
  apidocs.RecordingData:
    properties:
      data:
        $ref: '#/definitions/domain.Recording'
    type: object
...
  apidocs.SessionList:
    properties:
      sessions:
        items:
          $ref: '#/definitions/domain.PlaySession'
        type: array
      stats:
        $ref: '#/definitions/apidocs.SessionStats'
      total_count:
        type: integer
    type: object
...
  apidocs.StreamData:
    properties:
      data:
        $ref: '#/definitions/apidocs.StreamResponse'
    type: object
  apidocs.StreamList:
    properties:
      data:
        items:
          $ref: '#/definitions/apidocs.StreamResponse'
        type: array
      total:
        type: integer
    type: object
...
  apidocs.VODMountList:
    properties:
      data:
        items:
          $ref: '#/definitions/domain.VODMount'
        type: array
      total:
        type: integer
    type: object
...
  domain.GlobalConfig:
    properties:
      buffer:
        $ref: '#/definitions/config.BufferConfig'
      hooks:
        $ref: '#/definitions/config.HooksConfig'
      ingestor:
        $ref: '#/definitions/config.IngestorConfig'
      listeners:
        $ref: '#/definitions/config.ListenersConfig'
      log:
        $ref: '#/definitions/config.LogConfig'
      manager:
        $ref: '#/definitions/config.ManagerConfig'
      publisher:
        $ref: '#/definitions/config.PublisherConfig'
      server:
        $ref: '#/definitions/config.ServerConfig'
      sessions:
        $ref: '#/definitions/config.SessionsConfig'
      transcoder:
        $ref: '#/definitions/config.TranscoderConfig'
      watermarks:
        $ref: '#/definitions/config.WatermarksConfig'
    type: object
...
  domain.Hook:
    properties:
      enabled:
        type: boolean
      event_types:
        description: EventTypes filters which events trigger delivery. Empty = all
          events.
        items:
          $ref: '#/definitions/domain.EventType'
        type: array
      id:
        type: string
      max_retries:
        description: |-
          MaxRetries is the number of delivery attempts before giving up.
          0 means use the server default (3).
        type: integer
      metadata:
        additionalProperties:
          type: string
        description: |-
          Metadata holds user-defined key-value pairs merged into every event payload
          delivered by this hook. Useful for tagging events with custom context
          (e.g. environment, tenant ID, region) without modifying the server config.
        type: object
      name:
        type: string
      secret:
        description: HMAC-SHA256 signing secret (HTTP only)
        type: string
      stream_codes:
        allOf:
        - $ref: '#/definitions/domain.StreamCodeFilter'
        description: |-
          StreamCodes filters delivery by stream code.
          Only and Except are mutually exclusive; Only takes precedence when both are set.
          Omitting the field (nil) means all streams are included.
      target:
        description: HTTP URL or Kafka topic
        type: string
      timeout_sec:
        description: |-
          TimeoutSec is the per-attempt delivery timeout in seconds.
          0 means use the server default (10s).
        type: integer
      type:
        $ref: '#/definitions/domain.HookType'
    type: object
...
  handler.probeRequest:
    properties:
      ffmpeg_path:
        type: string
    type: object
...
  transcoder.ProbeResult:
    properties:
      encoders:
        additionalProperties:
          additionalProperties:
            type: boolean
          type: object
        type: object
      errors:
        items:
          type: string
        type: array
      muxers:
        additionalProperties:
          type: boolean
        type: object
      ok:
        type: boolean
      path:
        type: string
      version:
        type: string
      warnings:
        items:
          type: string
        type: array
    type: object
...
info:
  contact: {}
  description: REST API for live stream configuration, ingest/publish lifecycle, DVR
    recordings, and event hooks.
  title: Open Streamer API
  version: "1.0"
paths:
  /config:
    get:
      description: Returns available hardware accelerators (OS-detected), static enum
        lists, publisher listener ports, and the current runtime GlobalConfig.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apidocs.ConfigData'
      summary: Get server configuration.
      tags:
      - system
    post:
      consumes:
      - application/json
      description: Merges the request body into the current GlobalConfig and applies
        it — services are started or stopped to match.
      parameters:
      - description: Partial configuration to merge
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.GlobalConfig'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apidocs.ConfigUpdateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
      summary: Partially update server configuration.
      tags:
      - system
...
  /config/transcoder/probe:
    post:
      consumes:
      - application/json
      description: Inspects the binary at ffmpeg_path (empty = $PATH) and reports
        which required / optional encoders + muxers are available. Pure check — does
        not modify config.
      parameters:
      - description: Path to probe (empty = use $PATH)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.probeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/transcoder.ProbeResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
      summary: Probe an FFmpeg binary for app compatibility.
      tags:
      - system
  /config/yaml:
    get:
      description: Returns global_config, streams, and hooks bundled in a single YAML
        document. Pair with PUT /config/yaml to round-trip an editor.
      produces:
      - application/yaml
      responses:
        "200":
          description: YAML document
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get full system configuration as YAML
      tags:
      - system
    put:
      consumes:
      - application/yaml
      description: Replaces global_config, streams, and hooks with the request body.
        Validation errors are returned as a list. On success, the runtime manager
        + coordinator diff against the previous state and start/stop/restart services
        accordingly.
      parameters:
      - description: Full system configuration YAML document
        in: body
        name: body
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Replace full system configuration from YAML
      tags:
      - system
...
  /hooks:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apidocs.HookList'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
      summary: List hooks
      tags:
      - hooks
    post:
      consumes:
      - application/json
      parameters:
      - description: Hook configuration
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.Hook'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/apidocs.HookData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
      summary: Create hook
      tags:
      - hooks
  /hooks/{hid}:
    delete:
      parameters:
      - description: Hook ID
        in: path
        name: hid
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
      summary: Delete hook
      tags:
      - hooks
    get:
      parameters:
      - description: Hook ID
        in: path
        name: hid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apidocs.HookData'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
      summary: Get hook
      tags:
      - hooks
    put:
      consumes:
      - application/json
      parameters:
      - description: Hook ID
        in: path
        name: hid
        required: true
        type: string
      - description: Hook configuration
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.Hook'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apidocs.HookData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
      summary: Update hook
      tags:
      - hooks
...
  /recordings/{rid}:
    get:
      parameters:
      - description: Recording ID (= stream code)
        in: path
        name: rid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apidocs.RecordingData'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
      summary: Get recording
      tags:
      - recordings
  /recordings/{rid}/{file}:
    get:
      parameters:
      - description: Recording ID (= stream code)
        in: path
        name: rid
        required: true
        type: string
      - description: Segment filename (e.g. 000000.ts)
        in: path
        name: file
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
      summary: Serve DVR segment
      tags:
      - recordings
  /recordings/{rid}/info:
    get:
      parameters:
      - description: Recording ID (= stream code)
        in: path
        name: rid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
      summary: DVR recording info
      tags:
      - recordings
  /recordings/{rid}/playlist.m3u8:
    get:
      parameters:
      - description: Recording ID (= stream code)
        in: path
        name: rid
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: M3U8 body
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
      summary: Recording M3U8 playlist
      tags:
      - recordings
  /recordings/{rid}/timeshift.m3u8:
    get:
      parameters:
      - description: Recording ID (= stream code)
        in: path
        name: rid
        required: true
        type: string
      - description: Absolute start time (RFC3339)
        in: query
        name: from
        type: string
      - description: Relative start offset in seconds from recording start
        in: query
        name: offset_sec
        type: integer
      - description: 'Window duration in seconds (default: all remaining)'
        in: query
        name: duration
        type: integer
      produces:
      - text/plain
      responses:
        "200":
          description: M3U8 body
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
      summary: Timeshift VOD playlist
      tags:
      - recordings
  /sessions:
    get:
      parameters:
      - description: Filter by protocol (hls|dash|rtmp|srt|rtsp)
        in: query
        name: proto
        type: string
      - description: Filter by status (active|closed)
        in: query
        name: status
        type: string
      - description: Max sessions to return (default 0 = no cap)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apidocs.SessionList'
      summary: List play sessions
      tags:
      - sessions
  /sessions/{id}:
    delete:
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: kicked
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
      summary: Kick a play session
      tags:
      - sessions
    get:
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.PlaySession'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
      summary: Get a play session by ID
      tags:
      - sessions
  /streams:
    get:
      parameters:
      - description: Filter by status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apidocs.StreamList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
      summary: List streams
      tags:
      - streams
  /streams/{code}:
    delete:
      parameters:
      - description: Stream code
        in: path
        name: code
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
      summary: Delete stream
      tags:
      - streams
    get:
      parameters:
      - description: Stream code (a-zA-Z0-9_)
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apidocs.StreamData'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
      summary: Get stream
      tags:
      - streams
    post:
      consumes:
      - application/json
      parameters:
      - description: Stream code
        in: path
        name: code
        required: true
        type: string
      - description: Partial or full stream document
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/apidocs.StreamPutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apidocs.StreamData'
        "201":
          description: Created
          schema:
            $ref: '#/definitions/apidocs.StreamData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
      summary: Create or partial-update stream
      tags:
      - streams
...
  /vod:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apidocs.VODMountList'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
      summary: List VOD mounts
      tags:
      - vod
    post:
      consumes:
      - application/json
      parameters:
      - description: VOD mount
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.VODMount'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/apidocs.VODMountData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
      summary: Create VOD mount
      tags:
      - vod
  /vod/{name}:
    delete:
      parameters:
      - description: Mount name
        in: path
        name: name
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
      summary: Delete VOD mount
      tags:
      - vod
    get:
      parameters:
      - description: Mount name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apidocs.VODMountData'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
      summary: Get VOD mount
      tags:
      - vod
    put:
      consumes:
      - application/json
      parameters:
      - description: Mount name
        in: path
        name: name
        required: true
        type: string
      - description: VOD mount
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.VODMount'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apidocs.VODMountData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
      summary: Update VOD mount
      tags:
      - vod
...
  /watermarks:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apidocs.WatermarkAssetList'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
      summary: List watermark assets
      tags:
      - watermarks
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Image file (PNG / JPG / GIF, ≤ 8 MiB)
        in: formData
        name: file
        required: true
        type: file
      - description: Display name (defaults to filename)
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/apidocs.WatermarkAssetData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
        "413":
          description: Request Entity Too Large
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
        "415":
          description: Unsupported Media Type
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
      summary: Upload a watermark image
      tags:
      - watermarks
  /watermarks/{id}:
    delete:
      parameters:
      - description: Asset ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
      summary: Delete a watermark asset
      tags:
      - watermarks
    get:
      parameters:
      - description: Asset ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apidocs.WatermarkAssetData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
      summary: Get watermark metadata
      tags:
      - watermarks
...
swagger: "2.0"