# internal/api/server.go
// Package api implements the HTTP API server.
// All routes follow the REST conventions defined in api-conventions.mdc.
package api

import (
	"context"
	"fmt"
	"log/slog"
	"net/http"
	"path/filepath"
	"strings"
	"time"

	"github.com/go-chi/chi/v5"
	"github.com/go-chi/chi/v5/middleware"
	"github.com/go-chi/cors"
	_ "github.com/ntt0601zcoder/open-streamer/api/docs" // swag Register(SwaggerInfo)
	"github.com/ntt0601zcoder/open-streamer/config"
	"github.com/ntt0601zcoder/open-streamer/internal/api/handler"
	"github.com/ntt0601zcoder/open-streamer/internal/mediaserve"
	"github.com/ntt0601zcoder/open-streamer/internal/sessions"
	"github.com/prometheus/client_golang/prometheus/promhttp"
	"github.com/samber/do/v2"
)

// Server is the HTTP API server.
type Server struct {
	hlsDir  string
	dashDir string

	// Handler references stored for router rebuilding on config change.
	streamH    *handler.StreamHandler
	recordingH *handler.RecordingHandler
	hookH      *handler.HookHandler
	configH    *handler.ConfigHandler
	vodH       *handler.VODHandler
	sessionH   *handler.SessionHandler
	watermarkH *handler.WatermarkHandler

	// sessTracker is the shared play-sessions tracker. Used to wrap the
	// mediaserve mount with a tracking middleware so HLS / DASH segment
	// fetches are recorded as sessions. nil disables tracking entirely.
	sessTracker sessions.Tracker

	router *chi.Mux
	http   *http.Server
}

// New creates a Server and registers it with the DI injector.
// The server is constructed but not started — call StartWithConfig to begin accepting connections.
func New(i do.Injector) (*Server, error) {
	pub := do.MustInvoke[config.PublisherConfig](i)

	dashDir := strings.TrimSpace(pub.DASH.Dir)
	if dashDir == "" {
		dashDir = "./dash"
	}

	s := &Server{
		hlsDir:     pub.HLS.Dir,
		dashDir:    dashDir,
		streamH:    do.MustInvoke[*handler.StreamHandler](i),
		recordingH: do.MustInvoke[*handler.RecordingHandler](i),
		hookH:      do.MustInvoke[*handler.HookHandler](i),
		configH:    do.MustInvoke[*handler.ConfigHandler](i),
		vodH:       do.MustInvoke[*handler.VODHandler](i),
		sessionH:   do.MustInvoke[*handler.SessionHandler](i),
		watermarkH: do.MustInvoke[*handler.WatermarkHandler](i),
	}
	// Tracker is optional: only wrap mediaserve when the sessions feature is
	// wired in DI. Treats a missing provider as "feature disabled" rather
	// than a startup error.
	if t, err := do.Invoke[*sessions.Service](i); err == nil {
		s.sessTracker = t
	}
	return s, nil
}

func (s *Server) buildRouter(serverCfg *config.ServerConfig) *chi.Mux {
	r := chi.NewRouter()

	r.Use(middleware.RequestID)
	r.Use(middleware.RealIP)
	if serverCfg.CORS.Enabled {
		r.Use(cors.Handler(corsOptions(serverCfg.CORS)))
	}
	r.Use(middleware.Recoverer)
	r.Use(skipMediaLogger(slogAccessLogger))
	r.Use(middleware.Timeout(120 * time.Second))

	r.Get("/healthz", healthz)
	r.Get("/readyz", readyz)
	r.Handle("/metrics", promhttp.Handler())
	r.Get("/config", s.configH.GetConfig)
	r.Post("/config", s.configH.UpdateConfig)
	r.Get("/config/defaults", s.configH.GetConfigDefaults)
	r.Post("/config/transcoder/probe", s.configH.ProbeTranscoder)
	r.Get("/config/yaml", s.configH.GetConfigYAML)
	r.Put("/config/yaml", s.configH.ReplaceConfigYAML)

	r.Get("/swagger/doc.json", serveSwaggerJSON)
	r.Get("/swagger", http.RedirectHandler("/swagger/", http.StatusMovedPermanently).ServeHTTP)
	r.Get("/swagger/", serveSwaggerIndex)

	// Mount media routes under a sub-router so we can attach the play-sessions
	// middleware ONLY to /<code>/<segment> paths and not to the API endpoints
	// above. Chi composes Use+Mount cleanly here.
	r.Group(func(mr chi.Router) {
		if s.sessTracker != nil {
			mr.Use(sessions.HTTPMiddleware(s.sessTracker))
		}
		mediaserve.Mount(mr, s.hlsDir, s.dashDir)
	})

	r.Route("/streams", func(r chi.Router) {
		r.Get("/", s.streamH.List)
		r.Route("/{code}", func(r chi.Router) {
			r.Get("/", s.streamH.Get)
			r.Post("/", s.streamH.Put)
			r.Delete("/", s.streamH.Delete)
			r.Post("/restart", s.streamH.Restart)
			r.Post("/inputs/switch", s.streamH.SwitchInput)

			r.Get("/recordings", s.recordingH.ListByStream)
			r.Get("/sessions", s.sessionH.ListByStream)
		})
	})

	r.Route("/sessions", func(r chi.Router) {
		r.Get("/", s.sessionH.List)
		r.Route("/{id}", func(r chi.Router) {
			r.Get("/", s.sessionH.Get)
			r.Delete("/", s.sessionH.Delete)
		})
	})

	r.Route("/recordings/{rid}", func(r chi.Router) {
		r.Get("/", s.recordingH.Get)
		r.Get("/info", s.recordingH.Info)
		r.Get("/playlist.m3u8", s.recordingH.Playlist)
		r.Get("/timeshift.m3u8", s.recordingH.Timeshift)
		r.Get("/{file}", s.recordingH.ServeSegment)
	})

	r.Route("/hooks", func(r chi.Router) {
		r.Get("/", s.hookH.List)
		r.Post("/", s.hookH.Create)
		r.Route("/{hid}", func(r chi.Router) {
			r.Get("/", s.hookH.Get)
			r.Put("/", s.hookH.Update)
			r.Delete("/", s.hookH.Delete)
			r.Post("/test", s.hookH.Test)
		})
	})

	r.Route("/watermarks", func(r chi.Router) {
		r.Get("/", s.watermarkH.List)
		r.Post("/", s.watermarkH.Upload)
		r.Route("/{id}", func(r chi.Router) {
			r.Get("/", s.watermarkH.Get)
			r.Delete("/", s.watermarkH.Delete)
			r.Get("/raw", s.watermarkH.Raw)
		})
	})

	r.Route("/vod", func(r chi.Router) {
		r.Get("/", s.vodH.List)
		r.Post("/", s.vodH.Create)
		r.Route("/{name}", func(r chi.Router) {
			r.Get("/", s.vodH.Get)
			r.Put("/", s.vodH.Update)
			r.Delete("/", s.vodH.Delete)
			r.Get("/files", s.vodH.ListFiles)
			r.Post("/files", s.vodH.UploadFile)
			r.Delete("/files/*", s.vodH.DeleteFile)
			r.Get("/raw/*", s.vodH.Raw)
		})
	})

	return r
}

// slogAccessLogger is a chi-style access-log middleware that emits via slog
// at INFO level (one structured event per request). Replacing chi's
// middleware.Logger — which writes via the std `log` package and so
// bypasses slog level filtering — means setting `log.level: error` in the
// global config now actually silences these access lines as the operator
// expects. Format: "api: http request" with method/path/status/duration_ms
// and remote ip; chi's RequestID is included when present.
func slogAccessLogger(next http.Handler) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		// Skip the work entirely when slog will discard the line — saves
		// the ResponseWriter wrap + status capture on hot paths when the
		// operator runs at WARN/ERROR level.
		if !slog.Default().Enabled(r.Context(), slog.LevelInfo) {
			next.ServeHTTP(w, r)
			return
		}
		ww := middleware.NewWrapResponseWriter(w, r.ProtoMajor)
		start := time.Now()
		next.ServeHTTP(ww, r)
		slog.Info("api: http request",
			"method", r.Method,
			"path", r.URL.Path,
			"status", ww.Status(),
			"bytes", ww.BytesWritten(),
			"duration_ms", time.Since(start).Milliseconds(),
			"remote", r.RemoteAddr,
			"request_id", middleware.GetReqID(r.Context()),
		)
	})
}

// skipMediaLogger wraps a logging middleware so that requests for HLS/DASH
// media files (.m3u8, .ts, .m4s, .mpd, .mp4) bypass the logger entirely.
func skipMediaLogger(logger func(http.Handler) http.Handler) func(http.Handler) http.Handler {
	skip := map[string]bool{
		".m3u8": true,
		".ts":   true,
		".m4s":  true,
		".mpd":  true,
		".mp4":  true,
	}
	return func(next http.Handler) http.Handler {
		logged := logger(next)
		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
			if skip[strings.ToLower(filepath.Ext(r.URL.Path))] {
				next.ServeHTTP(w, r)
				return
			}
			logged.ServeHTTP(w, r)
		})
	}
}

func corsOptions(cfg config.CORSConfig) cors.Options {
	origins := cfg.AllowedOrigins
	if len(origins) == 0 {
		origins = []string{"*"}
	}
	methods := cfg.AllowedMethods
	if len(methods) == 0 {
		methods = []string{"GET", "POST", "PUT", "PATCH", "DELETE", "HEAD", "OPTIONS"}
	}
	headers := cfg.AllowedHeaders
	if len(headers) == 0 {
		headers = []string{"Accept", "Authorization", "Content-Type", "X-Request-ID"}
	}
	allowCredentials := cfg.AllowCredentials
	if allowCredentials {
		for _, o := range origins {
			if o == "*" {
				allowCredentials = false
				break
			}
		}
	}
	opts := cors.Options{
		AllowedOrigins:   origins,
		AllowedMethods:   methods,
		AllowedHeaders:   headers,
		ExposedHeaders:   cfg.ExposedHeaders,
		AllowCredentials: allowCredentials,
	}
	if cfg.MaxAge > 0 {
		opts.MaxAge = cfg.MaxAge
	}
	return opts
}

// StartWithConfig builds the router from the given ServerConfig and begins accepting
// HTTP connections. Blocks until ctx is cancelled.
// Called by the RuntimeManager each time the HTTP service is (re)started.
func (s *Server) StartWithConfig(ctx context.Context, cfg *config.ServerConfig) error {
	s.router = s.buildRouter(cfg)
	s.http = &http.Server{
		Addr:    cfg.HTTPAddr,
		Handler: s.router,
	}

	go func() {
		<-ctx.Done()
		shutCtx, cancel := context.WithTimeout(context.WithoutCancel(ctx), 10*time.Second)
		defer cancel()
		if err := s.http.Shutdown(shutCtx); err != nil {
			slog.Warn("api: HTTP shutdown", "err", err)
		}
	}()

	slog.Info("api: HTTP server listening", "addr", cfg.HTTPAddr)
	if err := s.http.ListenAndServe(); err != nil && err != http.ErrServerClosed {
		return fmt.Errorf("api: server: %w", err)
	}
	return nil
}

// healthz liveness probe.
// @Summary Liveness
// @Description Returns 200 if the process is up.
// @Tags system
// @Success 200 {string} string "ok"
// @Router /healthz [get].
func healthz(w http.ResponseWriter, _ *http.Request) {
	w.WriteHeader(http.StatusOK)
	_, _ = w.Write([]byte("ok"))
}

// readyz readiness probe.
// @Summary Readiness
// @Description Returns 200 when the server accepts traffic (basic check).
// @Tags system
// @Success 200 {string} string "ok"
// @Router /readyz [get].
func readyz(w http.ResponseWriter, r *http.Request) {
	healthz(w, r)
}